+++ CHANGES.md
@@ -9,6 +9,7 @@ libcups v3.0.0 (YYYY-MM-DD)
 - Fixed return values of `ippDateToTime` when the timezone isn't GMT.
 - Fixed a potential timing issue with `cupsEnumDests`.
 - Fixed a bug in the Avahi implementation of `cupsDNSSDBrowseNew`.
+- Fixed a memory leak in `httpClose`.
 
 
 libcups v3.0rc4 (2025-03-18)